move slice_interval function to utils
