+++ neuropacks/nsds_nwb/nsds_nwb.py
@@ -3,6 +3,8 @@ import scipy as sp
 
 from pynwb import NWBHDF5IO
 
+from neuropacks.nsds_nwb.utils import slice_interval
+
 
 class NSDSNWBAudio:
     def __init__(self, nwb_path):
@@ -207,29 +209,3 @@ class NSDSNWBAudio:
             fftd[freq > 0] *= 2
             self._stimulus_envelope = abs(sp.fft.ifft(fftd))
         return self._stimulus_envelope
-
-
-def slice_interval(start_time, stop_time, rate, t_offset=0):
-    '''
-    Returns a slice object.
-    '''
-    if start_time is not None and start_time < t_offset:
-        raise ValueError(f'start_time is before the timeseries onset, {t_offset}.')
-    if stop_time is not None and stop_time < t_offset:
-        raise ValueError(f'stop_time is before the timeseries onset, {t_offset}.')
-    if (start_time is not None and stop_time is not None) and start_time > stop_time:
-        raise ValueError('stop_time should be later than start_time.')
-
-    if start_time is not None:
-        dt_start = start_time - t_offset
-        i_start = int(dt_start * rate)
-    else:
-        i_start = None
-
-    if stop_time is not None:
-        dt_stop = stop_time - t_offset
-        i_stop = int(dt_stop * rate)
-    else:
-        i_stop = None
-
-    return slice(i_start, i_stop)
